+++ handlers/product-hub/helpers/constants.ts
@@ -1,0 +1,1 @@
+export const measureTime = true
