+++ handlers/product-hub/helpers/filterTableData.ts
@@ -1,0 +1,4 @@
+import { ProductHubItems } from '@prisma/client'
+
+// removing the ID cause it's not needed in the frontend
+export const filterTableData = ({ id, ...table }: ProductHubItems) => table
