+++ handlers/product-hub/helpers/getProtocolProducts.ts
@@ -1,0 +1,12 @@
+import { Protocol } from '@prisma/client'
+import { prisma } from 'server/prisma'
+
+export function getProtocolProducts(protocol: Protocol) {
+  return prisma.productHubItems.findMany({
+    where: {
+      protocol: {
+        equals: protocol,
+      },
+    },
+  })
+}
